Fix imports, split types, and improve logic
