+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveQuestion/InteractiveQuestion.tsx
@@ -33,11 +33,13 @@ export const _InteractiveQuestion = ({
       location={location}
       params={params}
       componentPlugins={plugins}
-      customTitle={customTitle}
-      withResetButton={withResetButton}
-      withTitle={withTitle}
     >
-      <InteractiveQuestionResult height={height} />
+      <InteractiveQuestionResult
+        height={height}
+        customTitle={customTitle}
+        withResetButton={withResetButton}
+        withTitle={withTitle}
+      />
     </InteractiveQuestionProvider>
   );
 };
