+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveQuestion/components/BackButton.tsx
@@ -1,5 +1,6 @@
-import { useInteractiveQuestionContext } from "embedding-sdk/components/public/InteractiveQuestion/context";
 import { DashboardBackButton } from "metabase/query_builder/components/view/ViewHeader/components";
+
+import { useInteractiveQuestionContext } from "../context";
 
 export const BackButton = () => {
   const { onNavigateBack } = useInteractiveQuestionContext();
