+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveQuestion/components/FilterButton.tsx
@@ -1,7 +1,8 @@
-import { useInteractiveQuestionData } from "embedding-sdk/components/public/InteractiveQuestion/context";
 import { FilterHeaderButton } from "metabase/query_builder/components/view/ViewHeader/components";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
+
+import { useInteractiveQuestionData } from "../hooks";
 
 export const FilterButton = ({ onClick }: { onClick: () => void }) => {
   const { question } = useInteractiveQuestionData();
