+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveQuestion/components/Title.tsx
@@ -1,15 +1,9 @@
 import { QuestionTitle } from "embedding-sdk/components/private/QuestionTitle";
-import {
-  useInteractiveQuestionContext,
-  useInteractiveQuestionData,
-} from "embedding-sdk/components/public/InteractiveQuestion/context";
+
+import { useInteractiveQuestionData } from "../hooks";
 
 export const Title = () => {
   const { question } = useInteractiveQuestionData();
-  const { customTitle, withTitle } = useInteractiveQuestionContext();
 
-  return (
-    withTitle &&
-    (customTitle || (question && <QuestionTitle question={question} />))
-  );
+  return question && <QuestionTitle question={question} />;
 };
