+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveQuestion/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-interactive-question-data";
+export * from "./use-load-question";
