+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveQuestion/index.ts
@@ -1,2 +1,3 @@
 export * from "./InteractiveQuestion";
 export * from "./components";
+export * from "./hooks";
